refactor(discovery): split devices.go (1269 → 4 files) by concern
Eighth worst-first file-size refactor. Function bodies relocated unchanged.

  devices.go         401  DeviceDiscovery + New + Start/Stop + accessors + name resolution
  devices_types.go   243  DiscoveredDevice + protocol info structs + enums + ComputeDisplayName
  devices_scan.go    464  Scan + per-protocol merge + dedupe + vendor/mDNS/display post-process
  devices_copy.go    197  deep-copy helpers for GetDevices snapshots

All under the 600-line warn budget.

Build, vet, lint (--new-from-rev=HEAD~1: 0 issues), discovery tests pass.
